Moved constants to closest scope
