+++ app/javascript/src/components/Dashboard/Contacts/Pane/CreateContact.jsx
@@ -2,8 +2,7 @@ import React from "react";
 
 import { Pane, Typography, Toastr } from "neetoui";
 
-import formInitialValues from "constants/formInitialValues";
-
+import { INITIAL_FORM_VALUES } from "./constants";
 import Form from "./Form";
 
 const CreateContact = ({ showPane, setShowPane }) => {
@@ -22,7 +21,7 @@ const CreateContact = ({ showPane, setShowPane }) => {
       </Pane.Header>
       <Form
         onClose={onClose}
-        contact={formInitialValues.contactsForm}
+        contact={INITIAL_FORM_VALUES}
         handleSubmit={handleCreate}
       />
     </Pane>
