+++ app/javascript/src/constants/formInitialValues.js
@@ -13,29 +13,4 @@ export default {
     password: "",
     passwordConfirmation: "",
   },
-  notesForm: {
-    title: "",
-    description: "",
-    contact: "",
-    tags: "",
-  },
-  editNotes: {
-    title: "How to claim the warranty?",
-    description:
-      '"Are you getting my texts???" she texted to him. He glanced at it and chuckled under his breath. Of course he was getting them, but if he wasn\'t getting',
-    contact: {
-      label: "User One",
-      value: "User One",
-    },
-    tags: {
-      label: "Getting Started",
-      value: "Getting Started",
-    },
-  },
-  contactsForm: {
-    firstName: "",
-    lastName: "",
-    email: "",
-    role: "",
-  },
 };
