+++ app/javascript/src/constants/formValidationSchemas.js
@@ -24,19 +24,4 @@ export default {
     lastName: yup.string().required("Required"),
     password: yup.string().required("Required"),
   }),
-  notesForm: yup.object().shape({
-    title: yup.string().required("Title is required"),
-    description: yup.string().required("Description is required"),
-    contact: yup.object().required("Assigned contact is required").nullable(),
-    tags: yup.object().required("Tags is required").nullable(),
-  }),
-  contactsForm: yup.object().shape({
-    firstName: yup.string().required("First Name is required"),
-    lastName: yup.string().required("Last Name is required"),
-    email: yup
-      .string()
-      .email("Invalid email address")
-      .required("Email is required"),
-    role: yup.object().required("Role is required").nullable(),
-  }),
 };
